Styling page and adding remove btn
