+++ scr/css/styles.css
@@ -15,6 +15,24 @@
   --large-font: 1.2em;
 }
 
+@keyframes shake {
+  0% {
+    transform: translateX(0);
+  }
+  25% {
+    transform: translateX(-5px);
+  }
+  50% {
+    transform: translateX(5px);
+  }
+  75% {
+    transform: translateX(-5px);
+  }
+  100% {
+    transform: translateX(0);
+  }
+}
+
 * {
   box-sizing: border-box;
 }
@@ -38,6 +56,11 @@ header {
   justify-content: space-between;
   align-items: center;
   align-items: center;
+}
+
+.hero {
+  padding: 0;
+  margin: 0;
 }
 
 .title-heading,
@@ -68,25 +91,8 @@ nav {
 }
 
 /* STYLES FOR THE RECIPE VIEW PAGE */
-
-.recipe-card {
-  text-align: center;
-}
-
-.recipe-card img {
-  display: block;
-  margin: auto;
-  padding: 0;
-}
-
 .recipe-card ul {
   text-align: left;
-  margin: 10px 100px;
-  color: var(--dark-grey);
-}
-
-.recipe-instructions {
-  padding: 10px 40px;
   color: var(--dark-grey);
 }
 
@@ -110,4 +116,66 @@ nav {
   box-sizing: border-box;
   margin-bottom: 8px;
   text-align: center;
+}
+
+/* FAVORITE VIEW STYLES */
+.featured-section,
+.recipes-section,
+.favorites-section {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 16px;
+  justify-content: center;
+}
+
+.recipe-card {
+  position: relative;
+  border: 1px solid #ddd;
+  border-radius: 8px;
+  padding: 16px;
+  max-width: 300px;
+  text-align: center;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+.recipe-card h3 {
+  margin-top: 8px;
+  font-size: 1.2em;
+}
+
+.recipe-card img {
+  max-width: 100%;
+  height: auto;
+  border-radius: 4px;
+  margin: 8px 0;
+}
+
+.recipe-instructions {
+  font-size: 0.9em;
+  color: #555;
+}
+
+/* REMOVE BTTN STYLES */
+.remove-button {
+  position: absolute;
+  top: 8px;
+  right: 8px;
+  background-color: #ff5c5c;
+  color: white;
+  border: none;
+  border-radius: 50%;
+  width: 24px;
+  height: 24px;
+  cursor: pointer;
+  font-size: 1em;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  padding: 0;
+}
+
+/* ADD RECIPE BTTN STYLES */
+.favorite-btn.shake {
+  animation: shake 0.5s ease;
+  color: red;
 }
