+++ scr/js/APIHandler.js
@@ -1,5 +1,3 @@
-// API Handler
-
 export async function getFeaturedRecipes() {
   const response = await fetch(
     'https://www.themealdb.com/api/json/v1/1/filter.php?c=Dessert'
