+++ scr/js/utils.js
@@ -4,7 +4,7 @@ export function createElement(type, props = {}, children = []) {
   //props: {textContent: "Hello World", id: "header1", "data-productId": 123, ...}
   Object.entries(props).forEach(([key, value]) => {
     if (~key.indexOf('-')) {
-      element.setAttribute(key, value); //add attributes
+      element.setAttribute(key, value);
     } else {
       element[key] = value;
     }
